+++ src/main/kotlin/de/pause/features/user/data/repo/UserRepository.kt
@@ -1,6 +1,11 @@
-package de.pause.model
+package de.pause.features.user.data.repo
 
-import de.pause.db.*
+import de.pause.database.suspendTransaction
+import de.pause.features.user.data.dao.*
+import de.pause.features.user.data.dto.LoginRequest
+import de.pause.features.user.data.dto.RegisterRequest
+import de.pause.features.user.data.dto.UserDto
+import de.pause.features.user.data.dto.UserPrincipal
 import org.jetbrains.exposed.sql.Op
 import org.jetbrains.exposed.sql.SizedCollection
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
